+++ pkg/symbol/addr2line/symtab.go
@@ -1,4 +1,4 @@
-// Copyright 2022 The Parca Authors
+// Copyright 2022-2023 The Parca Authors
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -15,26 +15,31 @@
 package addr2line
 
 import (
+	"bytes"
 	"debug/elf"
-	"errors"
+	"encoding/binary"
 	"fmt"
-	"sort"
+	"io"
+	"strings"
 
 	"github.com/go-kit/log"
-	"github.com/go-kit/log/level"
 
 	pb "github.com/parca-dev/parca/gen/proto/go/parca/metastore/v1alpha1"
 	"github.com/parca-dev/parca/pkg/profile"
 	"github.com/parca-dev/parca/pkg/symbol/demangle"
+	"github.com/parca-dev/parca/pkg/symbol/symbolsearcher"
+)
+
+const (
+	pltSuffix = "@plt" // add pltSuffix for plt symbol to keep consistent with perf
 )
 
 // SymtabLiner is a liner which utilizes .symtab and .dynsym sections.
 type SymtabLiner struct {
 	logger log.Logger
 
-	// symbols contains sorted symbols.
-	symbols   []elf.Symbol
 	demangler *demangle.Demangler
+	searcher  symbolsearcher.Searcher
 
 	filename string
 	f        *elf.File
@@ -47,9 +52,10 @@ func Symbols(logger log.Logger, filename string, f *elf.File, demangler *demangl
 		return nil, fmt.Errorf("failed to fetch symbols from object file: %w", err)
 	}
 
+	searcher := symbolsearcher.New(symbols)
 	return &SymtabLiner{
 		logger:    log.With(logger, "liner", "symtab"),
-		symbols:   symbols,
+		searcher:  searcher,
 		demangler: demangler,
 		filename:  filename,
 		f:         f,
@@ -65,25 +71,14 @@ func (lnr *SymtabLiner) File() string {
 }
 
 func (lnr *SymtabLiner) PCRange() ([2]uint64, error) {
-	if len(lnr.symbols) == 0 {
-		return [2]uint64{}, errors.New("no symbols found")
-	}
-
-	return [2]uint64{
-		lnr.symbols[0].Value,
-		lnr.symbols[len(lnr.symbols)-1].Value + lnr.symbols[len(lnr.symbols)-1].Size,
-	}, nil
+	return lnr.searcher.PCRange()
 }
 
 // PCToLines looks up the line number information for a program counter (memory address).
 func (lnr *SymtabLiner) PCToLines(addr uint64) (lines []profile.LocationLine, err error) {
-	i := sort.Search(len(lnr.symbols), func(i int) bool {
-		return lnr.symbols[i].Value > addr
-	})
-
-	if i < 1 || i > len(lnr.symbols) {
-		level.Debug(lnr.logger).Log("msg", "failed to find symbol for address", "addr", addr)
-		return nil, errors.New("failed to find symbol for address")
+	name, err := lnr.searcher.Search(addr)
+	if err != nil {
+		return nil, err
 	}
 
 	var (
@@ -91,12 +86,20 @@ func (lnr *SymtabLiner) PCToLines(addr uint64) (lines []profile.LocationLine, er
 		line int64 // 0
 	)
 
+	// plt symbol suffix with pltSuffix
+	// to demangle name, we should remove the pltSuffix first
+	// and then add it to demangled name
+	isplt := strings.HasSuffix(name, pltSuffix)
+	result := lnr.demangler.Demangle(&pb.Function{
+		SystemName: strings.TrimSuffix(name, pltSuffix),
+		Filename:   file,
+	})
+	if isplt {
+		result.Name = result.Name + pltSuffix
+	}
 	lines = append(lines, profile.LocationLine{
-		Line: line,
-		Function: lnr.demangler.Demangle(&pb.Function{
-			SystemName: lnr.symbols[i-1].Name,
-			Filename:   file,
-		}),
+		Line:     line,
+		Function: result,
 	})
 	return lines, nil
 }
@@ -106,14 +109,54 @@ func (lnr *SymtabLiner) PCToLines(addr uint64) (lines []profile.LocationLine, er
 // to facilitate searching.
 func symtab(objFile *elf.File) ([]elf.Symbol, error) {
 	syms, sErr := objFile.Symbols()
+	dynSyms, dErr := objFile.DynamicSymbols()
 
-	if sErr != nil {
+	var pltSymbols []elf.Symbol
+	// see symbol-elf.c/dso__synthesize_plt_symbols
+	if pltRelSection, pltSection := objFile.Section(".rela.plt"), objFile.Section(".plt"); dErr == nil &&
+		pltRelSection != nil && pltSection != nil &&
+		objFile.Sections[pltRelSection.Link].Type == elf.SHT_DYNSYM {
+		data, err := io.ReadAll(pltRelSection.Open())
+		if err != nil {
+			return nil, fmt.Errorf("failed to data of .rela.plt section:%s", err)
+		}
+		var rela elf.Rela64
+		b := bytes.NewReader(data)
+
+		// in perf script, it use pltSection.Offset
+		// but the computeBase return symbol address, which require pltSection.Addr
+		// - ET_EXEC, pltSection.Addr and  pltSection.Offset not same
+		// - ET_DYNSYM, pltSection.Addr and  pltSection.Offset is same
+		off := pltSection.Addr
+		for b.Len() > 0 {
+			off += pltSection.Entsize
+			err := binary.Read(b, objFile.ByteOrder, &rela)
+			if err != nil {
+				return nil, fmt.Errorf("read plt section error, err:%s, section:%v", err, *pltRelSection)
+			}
+			// see applyRelocationsAMD64 go1.19.3/src/debug/elf/file.go:664
+			i := rela.Info >> 32
+			if i > uint64(len(dynSyms)) || i == 0 {
+				continue
+			}
+			s := dynSyms[i-1]
+
+			pltSymbols = append(pltSymbols, elf.Symbol{
+				Name:    s.Name + pltSuffix,
+				Info:    elf.ST_INFO(elf.STB_GLOBAL, elf.STT_FUNC),
+				Section: elf.SectionIndex(1), // just to pass elfSymIsFunction's section check
+				Value:   off,
+				Size:    pltRelSection.Entsize,
+				Version: "",
+				Library: "",
+			})
+		}
+	}
+
+	if sErr != nil && dErr != nil {
 		return nil, fmt.Errorf("failed to read symbol sections: %w", sErr)
 	}
 
-	sort.SliceStable(syms, func(i, j int) bool {
-		return syms[i].Value < syms[j].Value
-	})
-
+	syms = append(syms, append(dynSyms, pltSymbols...)...)
 	return syms, nil
 }