+++ ui/packages/shared/profile/src/ProfileMetricsGraph/index.tsx
@@ -15,11 +15,12 @@ import {useEffect, useState} from 'react';
 
 import {RpcError} from '@protobuf-ts/runtime-rpc';
 
-import {Label, QueryRangeResponse, QueryServiceClient, Timestamp} from '@parca/client';
+import {Duration, Label, QueryRangeResponse, QueryServiceClient, Timestamp} from '@parca/client';
 import {DateTimeRange, useGrpcMetadata, useParcaContext} from '@parca/components';
+import {getStepDuration} from '@parca/functions';
 import {Query} from '@parca/parser';
 
-import {ProfileSelection, SingleProfileSelection} from '..';
+import {MergedProfileSelection, ProfileSelection} from '..';
 import MetricsGraph from '../MetricsGraph';
 import useDelayedLoader from '../useDelayedLoader';
 
@@ -29,9 +30,9 @@ interface ProfileMetricsGraphProps {
   profile: ProfileSelection | null;
   from: number;
   to: number;
-  select: (source: ProfileSelection) => void;
   setTimeRange: (range: DateTimeRange) => void;
   addLabelMatcher: (key: string, value: string) => void;
+  onPointClick: (timestamp: number, labels: Label[], queryExpression: string) => void;
 }
 
 export interface IQueryRangeState {
@@ -61,11 +62,13 @@ export const useQueryRange = (
         error: null,
       });
 
+      const stepDuration = getStepDuration(start, end);
       const call = client.queryRange(
         {
           query: queryExpression,
           start: Timestamp.fromDate(new Date(start)),
           end: Timestamp.fromDate(new Date(end)),
+          step: Duration.create(stepDuration),
           limit: 0,
         },
         {meta: metadata}
@@ -86,9 +89,9 @@ const ProfileMetricsGraph = ({
   profile,
   from,
   to,
-  select,
   setTimeRange,
   addLabelMatcher,
+  onPointClick,
 }: ProfileMetricsGraphProps): JSX.Element => {
   const {isLoading, response, error} = useQueryRange(queryClient, queryExpression, from, to);
   const isLoaderVisible = useDelayedLoader(isLoading);
@@ -119,9 +122,7 @@ const ProfileMetricsGraph = ({
   const series = response?.series;
   if (series !== null && series !== undefined && series?.length > 0) {
     const handleSampleClick = (timestamp: number, _value: number, labels: Label[]): void => {
-      select(
-        new SingleProfileSelection(Query.parse(queryExpression).profileName(), labels, timestamp)
-      );
+      onPointClick(timestamp, labels, queryExpression);
     };
 
     return (
@@ -133,7 +134,7 @@ const ProfileMetricsGraph = ({
           data={series}
           from={from}
           to={to}
-          profile={profile as SingleProfileSelection}
+          profile={profile as MergedProfileSelection}
           setTimeRange={setTimeRange}
           onSampleClick={handleSampleClick}
           onLabelClick={addLabelMatcher}