+++ ui/packages/shared/profile/src/ProfileSelector/index.tsx
@@ -24,22 +24,23 @@ import {
   DateTimeRangePicker,
   useGrpcMetadata,
 } from '@parca/components';
+import {getStepDuration, getStepDurationInMilliseconds} from '@parca/functions';
 import {CloseIcon} from '@parca/icons';
 import {Query} from '@parca/parser';
 
-import {ProfileSelection} from '..';
+import {MergedProfileSelection, ProfileSelection} from '..';
 import MatchersInput from '../MatchersInput/index';
 import ProfileMetricsGraph from '../ProfileMetricsGraph';
 import ProfileTypeSelector from '../ProfileTypeSelector/index';
 import CompareButton from './CompareButton';
-import MergeButton from './MergeButton';
 
 export interface QuerySelection {
   expression: string;
   from: number;
   to: number;
-  merge: boolean;
   timeSelection: string;
+  mergeFrom?: number;
+  mergeTo?: number;
 }
 
 interface ProfileSelectorProps {
@@ -125,18 +126,29 @@ const ProfileSelector = ({
     enforcedProfileName !== '' ? enforcedProfileNameQuery() : Query.parse(queryExpressionString);
   const selectedProfileName = query.profileName();
 
-  const setNewQueryExpression = (expr: string, merge: boolean): void => {
+  const setNewQueryExpression = (expr: string): void => {
+    const query = enforcedProfileName !== '' ? enforcedProfileNameQuery() : Query.parse(expr);
+    const delta = query.profileType().delta;
+    const from = timeRangeSelection.getFromMs();
+    const to = timeRangeSelection.getToMs();
+    const mergeParams = delta
+      ? {
+          mergeFrom: from,
+          mergeTo: to,
+        }
+      : {};
+
     selectQuery({
       expression: expr,
-      from: timeRangeSelection.getFromMs(),
-      to: timeRangeSelection.getToMs(),
-      merge,
+      from,
+      to,
       timeSelection: timeRangeSelection.getRangeKey(),
+      ...mergeParams,
     });
   };
 
   const setQueryExpression = (): void => {
-    setNewQueryExpression(query.toString(), false);
+    setNewQueryExpression(query.toString());
   };
 
   const addLabelMatcher = (key: string, value: string): void => {
@@ -145,12 +157,8 @@ const ProfileSelector = ({
     const newValue = value.includes('\\') ? value.replaceAll('\\', '\\\\') : value;
     const [newQuery, changed] = Query.parse(queryExpressionString).setMatcher(key, newValue);
     if (changed) {
-      setNewQueryExpression(newQuery.toString(), false);
+      setNewQueryExpression(newQuery.toString());
     }
-  };
-
-  const setMergedSelection = (): void => {
-    setNewQueryExpression(queryExpressionString, true);
   };
 
   const setMatchersString = (matchers: string): void => {
@@ -176,7 +184,6 @@ const ProfileSelector = ({
     queryExpressionString === '' ||
     queryExpressionString === '{}';
 
-  const mergeDisabled = selectedProfileName === '' || querySelection.expression === undefined;
   const compareDisabled = selectedProfileName === '' || querySelection.expression === undefined;
 
   return (
@@ -207,7 +214,6 @@ const ProfileSelector = ({
           <ButtonGroup>
             {!searchDisabled && (
               <>
-                <MergeButton disabled={mergeDisabled} onClick={setMergedSelection} />
                 {!comparing && (
                   <CompareButton disabled={compareDisabled} onClick={handleCompareClick} />
                 )}
@@ -232,19 +238,17 @@ const ProfileSelector = ({
           )}
         </div>
       </Card.Header>
-      {!querySelection.merge && (
+      {
         <Card.Body>
           {querySelection.expression !== undefined &&
           querySelection.expression.length > 0 &&
           querySelection.from !== undefined &&
-          querySelection.to !== undefined &&
-          (profileSelection == null || profileSelection.Type() !== 'merge') ? (
+          querySelection.to !== undefined ? (
             <ProfileMetricsGraph
               queryClient={queryClient}
               queryExpression={querySelection.expression}
               from={querySelection.from}
               to={querySelection.to}
-              select={selectProfile}
               profile={profileSelection}
               setTimeRange={(range: DateTimeRange) => {
                 setTimeRangeSelection(range);
@@ -252,15 +256,24 @@ const ProfileSelector = ({
                   expression: queryExpressionString,
                   from: range.getFromMs(),
                   to: range.getToMs(),
-                  merge: false,
                   timeSelection: range.getRangeKey(),
                 });
               }}
               addLabelMatcher={addLabelMatcher}
+              onPointClick={(timestamp, labels, queryExpression) => {
+                const stepDuration = getStepDuration(querySelection.from, querySelection.to);
+                const stepDurationInMilliseconds = getStepDurationInMilliseconds(stepDuration);
+                const isDeltaType = Query.parse(queryExpression).profileType().delta;
+                const mergeFrom = timestamp;
+                const mergeTo = isDeltaType ? mergeFrom + stepDurationInMilliseconds : mergeFrom;
+                selectProfile(
+                  new MergedProfileSelection(mergeFrom, mergeTo, labels, queryExpression)
+                );
+              }}
             />
           ) : (
             <>
-              {(profileSelection == null || profileSelection.Type() !== 'merge') && (
+              {profileSelection == null && (
                 <div className="my-20 text-center">
                   <p>Run a query, and the result will be displayed here.</p>
                 </div>
@@ -268,7 +281,7 @@ const ProfileSelector = ({
             </>
           )}
         </Card.Body>
-      )}
+      }
     </Card>
   );
 };